Add rudimentary support for ALTER DEFAULT PRIVILGES
Caveat: Postgres default privileges are a little weird in that they are
owned by a role, and they will only apply if that role is the one
creating the new object. For now I'm assuming "postgres", so you need to
make sure that postgres is the user owning these newly objects that you
want to have the default privileges copied into.

Really, default privileges are a shortcut that you shouldn't need if you
use this grants tool.